requirements.txt
pyham_pe
msgpack
pyham_ax25
ZODB
BTrees
transaction
ZEO
podman
click
tabulate
fastapi
uvicorn[standard]
jinja2
python-multipart
argon2-cffi
pydantic
pydantic_settings